Split Web App and database run stages
